Fix: Clear local asteroids only if remote data is not empty.

+++ app/src/main/java/com/udacity/asteroidradar/features/mainscreen/data/datasource/local/AsteroidsFeedLocalDataSourceImpl.kt
@@ -21,11 +21,13 @@ class AsteroidsFeedLocalDataSourceImpl(
     }
 
     override suspend fun saveFeedToLocalDatabase(asteroidsFeed: List<AsteroidsFeedItem>) {
-        asteroidDao.deleteOldAsteroids(currentDate = getCurrentDate())
-        asteroidsFeed.forEach {
-            it.createdAt = getCurrentDate()
-            it.modifiedAt = getCurrentDate()
-            asteroidDao.insert(it)
+        if (asteroidsFeed.isNotEmpty()) {
+            asteroidDao.deleteOldAsteroids(currentDate = getCurrentDate())
+            asteroidsFeed.forEach {
+                it.createdAt = getCurrentDate()
+                it.modifiedAt = getCurrentDate()
+                asteroidDao.insert(it)
+            }
         }
     }
 }
